requirements.txt
JPype1==0.6.3
mamba==0.9.3
